-Make comments commponent work with custom reprots

+++ src/components/report-comment/report-comment.ts
@@ -1,9 +1,9 @@
-import {AfterViewInit, Component, Input, OnInit, ViewChild} from '@angular/core';
+import {AfterViewChecked, Component, Input, OnInit, ViewChild} from '@angular/core';
 import {ReportCommentProvider} from "../../providers/report-comment/report-comment";
 import {ReportComment} from "../../models/reportComment";
 import {Student} from "../../models";
 import {AccountService} from "../../services/account";
-import {ActionSheetController} from "ionic-angular";
+import {ActionSheetController, ToastController} from "ionic-angular";
 
 /**
  * Generated class for the ReportCommentComponent component.
@@ -15,53 +15,71 @@ import {ActionSheetController} from "ionic-angular";
   selector: 'report-comment',
   templateUrl: 'report-comment.html'
 })
-export class ReportCommentComponent implements OnInit, AfterViewInit {
+export class ReportCommentComponent implements OnInit, AfterViewChecked {
   @Input() date:string;
   @Input() student:Student;
-  @Input() reportId:string;
-  public inEditModeComments: object = {};
+  @Input() reportId: number;
+
   public shouldShowComments:boolean = false;
   public canSubmitComment:boolean = false;
+  public canDeleteComment: boolean = false;
+  public canEditComment: boolean = false;
+
   public newCommentToBeSubmitted:string ="";
   public currentReportComments:ReportComment[];
+
   public isCommentsSectionExpanded:boolean = false;
   public isSendingComment:boolean = false;
   public isEditedCommentsLoading: object = {};
+  public inEditModeComments: object = {};
+  public isLoadingComments: boolean = false;
+  private shouldScrollToBottom: boolean = true;
   @ViewChild('commentsContainer') private commentsContainer: any;
 
-  constructor(private commentsProvider: ReportCommentProvider, private accountService: AccountService,
-              public actionSheetCtrl: ActionSheetController) {
+  constructor(private commentsProvider: ReportCommentProvider, public accountService: AccountService,
+              public actionSheetCtrl: ActionSheetController, private toastCtrl: ToastController) {
+    if (this.reportId == undefined || this.reportId == null) {
       this.shouldShowComments = this.accountService.getUserRole().dailyReportCommentView;
       this.canSubmitComment =  this.accountService.getUserRole().dailyReportCommentCreate;
+      this.canDeleteComment = this.accountService.getUserRole().dailyReportCommentDelete;
+      this.canEditComment = this.accountService.getUserRole().dailyReportCommentEdit;
+    } else {
+      this.shouldShowComments = this.accountService.getUserRole().reportCommentView;
+      this.canSubmitComment = this.accountService.getUserRole().reportCommentCreate;
+      this.canDeleteComment = this.accountService.getUserRole().dailyReportCommentDelete;
+      this.canEditComment = this.accountService.getUserRole().dailyReportCommentEdit;
+    }
+
   }
 
   ngOnInit() {
-    // this.scrollToBottom();
+    this.scrollToBottom();
   }
-
-  ngAfterViewInit() {
-    // this.scrollToBottom();
+ 
+  ngAfterViewChecked() {
+    this.scrollToBottom();
   }
 
   scrollToBottom(): void {
-    try {
-      this.commentsContainer.nativeElement.scrollTop = this.commentsContainer.nativeElement.scrollHeight;
-    } catch(err) { }
+    if (this.shouldScrollToBottom) {
+      try {
+        this.commentsContainer.nativeElement.scrollTop = this.commentsContainer.nativeElement.scrollHeight;
+      } catch (err) {
+      }
+    }
   }
 
   toggleCommentsSection(){
     this.isCommentsSectionExpanded = !this.isCommentsSectionExpanded;
     if(this.isCommentsSectionExpanded){
+      this.isLoadingComments = true;
       this.commentsProvider.getComments(this.date,this.student.studentId,this.reportId)
         .subscribe(comments =>{
+          this.shouldScrollToBottom = true;
+          this.isLoadingComments = false;
           this.currentReportComments = comments;
-          this.commentsContainer.changes.subscribe(() => {
-            const containerElement: HTMLElement = this.commentsContainer.nativeElement;
-            containerElement.lastElementChild.scrollIntoView(true);
-          });
-
         },error1 => {
-          console.log("errr");
+          this.showErrorToast("retrieve report comments");
         })
     }
 
@@ -72,29 +90,32 @@ export class ReportCommentComponent implements OnInit, AfterViewInit {
     this.commentsProvider.postNewComment(this.date,this.student.studentId,
       this.newCommentToBeSubmitted.trim(), this.reportId)
       .subscribe(newlySubmittedComment => {
+        this.shouldScrollToBottom = true;
         this.newCommentToBeSubmitted = "";
         this.isSendingComment = false;
         this.currentReportComments.push(newlySubmittedComment)
       }, error1 => {
         this.isSendingComment = false;
-        console.log("errr");
+        this.showErrorToast("submit new comment");
       });
   }
 
 
   onDeleteComment(comment: ReportComment, index: number) {
+    this.shouldScrollToBottom = false;
     this.isEditedCommentsLoading[comment.id] = true;
-    this.commentsProvider.deleteComment(comment.id)
+    this.commentsProvider.deleteComment(comment.id, this.reportId)
       .subscribe(() => {
         this.isEditedCommentsLoading[comment.id] = false;
         this.currentReportComments.splice(index, 1);
       }, error1 => {
         this.isEditedCommentsLoading[comment.id] = false;
-        console.log("errr");
+        this.showErrorToast("delete comment");
       })
   }
 
   toggleEditMode(comment: ReportComment) {
+    this.shouldScrollToBottom = false;
     if (this.inEditModeComments[comment.id]) {
       this.isEditedCommentsLoading[comment.id] = true;
       this.commentsProvider.editComment(this.date, this.student.studentId,
@@ -105,7 +126,7 @@ export class ReportCommentComponent implements OnInit, AfterViewInit {
         }, error1 => {
           this.isEditedCommentsLoading[comment.id] = false;
           this.inEditModeComments[comment.id] = false;
-          console.log("errr");
+          this.showErrorToast("edit comment");
         });
     } else {
       this.isEditedCommentsLoading[comment.id] = false;
@@ -117,29 +138,34 @@ export class ReportCommentComponent implements OnInit, AfterViewInit {
     let actionSheet = this.actionSheetCtrl.create({
       title: 'Comment options',
       enableBackdropDismiss: true,
-      buttons: [
-        {
-          text: 'Edit',
-          icon: "ios-create-outline",
-          handler: () => {
-            this.toggleEditMode(comment);
-          }
-        },
-        {
-          text: 'Delete',
-          icon: "ios-trash-outline",
-          role: 'destructive',
-          handler: () => {
-            this.onDeleteComment(comment, index)
-          }
-        }
-      ]
     });
-
+    if (this.canEditComment) {
+      actionSheet.addButton({
+        text: 'Edit',
+        icon: "ios-create-outline",
+        handler: () => {
+          this.toggleEditMode(comment);
+        }
+      });
+    }
+    if (this.canDeleteComment) {
+      actionSheet.addButton({
+        text: 'Delete',
+        icon: "ios-trash-outline",
+        role: 'destructive',
+        handler: () => {
+          this.onDeleteComment(comment, index)
+        }
+      })
+    }
     actionSheet.present();
   }
 
-  trimComment(commentText: string) {
-    return commentText.trim();
+  showErrorToast(operationName: string) {
+    let toast = this.toastCtrl.create({
+      message: 'Failed to ' + operationName,
+      duration: 3000,
+    });
+    toast.present();
   }
 }
